requirements.txt
pandas
numpy
scikit-learn
streamlit
plotly
xgboost
joblibpip install -r requirements.txt